+++ tests/integration/cases/call_as_expression/Erlang_call.erl
@@ -1,0 +1,9 @@
+-module(check).
+-export([x/0]).
+x() ->
+    process(_, _) -> ok.
+    Items = [
+    process(1, 42),
+    process(2, 100)
+    ],
+    Items.