+++ tests/integration/cases/call_as_expression/Java_call.java
@@ -1,0 +1,9 @@
+class Check {
+static Object process(Object... args) { return null; }
+    public static void check() {
+var items = new Object[]{
+process(1, 42),
+process(2, 100),
+};
+    }
+}