+++ tests/integration/cases/call_as_expression/Julia_call.jl
@@ -1,0 +1,5 @@
+process(args...; kwargs...) = nothing
+items = [
+process(a=1, b=42),
+process(a=2, b=100),
+]