+++ tests/integration/cases/call_as_expression/Perl_call.pl
@@ -1,0 +1,5 @@
+sub process {}
+my $items = [
+process(1, 42),
+process(2, 100),
+];