+++ tests/integration/cases/call_as_expression/Ruby_call.rb
@@ -1,0 +1,5 @@
+def process(*a); end
+items = [
+process(a: 1, b: 42),
+process(a: 2, b: 100),
+]