+++ tests/integration/cases/call_as_expression/Scala_call.scala
@@ -1,0 +1,7 @@
+object Check {
+def process(a: Any = null, b: Any = null): Any = null
+val items = List(
+process(a = 1, b = 42),
+process(a = 2, b = 100),
+)
+}